Структуры перенесены в отдельный файл StructDescription
